Migrate to TS

+++ src/commands/puzzle/index.ts
@@ -1,15 +1,16 @@
-const { SlashCommandBuilder, ChannelType, ButtonStyle, ActionRowBuilder, ButtonBuilder, AttachmentBuilder, EmbedBuilder } = require('discord.js');
-const { chromium } = require("playwright");
-const slugify = require("slugify");
-const { ACTIVE_PUZZLES_CHANNEL_CATEGORY_ID, PUZZ_WATCHERS_ROLE_ID, OLD_PUZZLES_CHANNEL_CATEGORY_ID } = require('../../config');
-const dayjs = require('dayjs');
-const relativeTime = require('dayjs/plugin/relativeTime');
+import { SlashCommandBuilder, ChannelType, ButtonStyle, ActionRowBuilder, ButtonBuilder, AttachmentBuilder, EmbedBuilder, ChatInputCommandInteraction, GuildChannel, CategoryChannel, Collection, NonThreadGuildBasedChannel, TextChannel } from 'discord.js';
+import { SlashCommand } from '../'
+import { chromium } from "playwright";
+import slugify from "slugify";
+import { ACTIVE_PUZZLES_CHANNEL_CATEGORY_ID, PUZZ_WATCHERS_ROLE_ID, OLD_PUZZLES_CHANNEL_CATEGORY_ID } from '../../config';
+import dayjs from 'dayjs';
+import relativeTime from 'dayjs/plugin/relativeTime';
 dayjs.extend(relativeTime);
 
 const NEW_PUZZLE_SUBCMD = "new";
 const CLOSE_PUZZLE_SUBCMD = "close";
 
-module.exports = {
+export default {
 	data: new SlashCommandBuilder()
 		.setName('puzzle')
 		.setDescription("Manage puzzles")
@@ -29,7 +30,7 @@ module.exports = {
 				.setName(CLOSE_PUZZLE_SUBCMD)
 				.setDescription('Finishes and archives a puzzle')
 		),
-	async execute(interaction) {
+	async execute(interaction: ChatInputCommandInteraction) {
 		switch (interaction.options.getSubcommand()) {
 			case NEW_PUZZLE_SUBCMD:
 				// do new puzzle stuff
@@ -41,31 +42,36 @@ module.exports = {
 				break;
 		}
 	}
-};
+} as SlashCommand;
 
-async function newpuzzle(interaction) {
+async function newpuzzle(interaction: ChatInputCommandInteraction) {
 	try {
-		const DFAC_REGEX = /https:\/\/downforacross.com\/beta\/game\/(.*)/im //create this each function so that .test() doesn't mess up lastindex
 		await interaction.deferReply({ephemeral: true});
-		const url = interaction.options.getString('url');
 
-		const match = url.match(DFAC_REGEX)
-		if (!match) {
-			await interaction.editReply({
-				content: "⚠️ Your URL didn't seem to be quite what I expected.\nMake sure it's got \"/beta/game\" in it to ensure it's a puzzle session!",
-				ephemeral: true
-			})
+		// Ensure the `guild` property is accessible
+		const guild = interaction.guild;
+		if (!guild) {
+			await interaction.editReply("⚠️ I couldn't find the guild for this interaction!")
+			return
+		}
+
+		const DFAC_REGEX = /https:\/\/downforacross.com\/beta\/game\/(.*)/im //create this each function so that .test() doesn't mess up lastindex
+		const url = interaction.options.getString('url');
+		const match = url?.match(DFAC_REGEX)
+		if (!match || !url) {
+			await interaction.editReply("⚠️ Your URL didn't seem to be quite what I expected.\nMake sure it's got \"/beta/game\" in it to ensure it's a puzzle session!")
 			return
 		}
 
 		// Check to see if we have too many puzzles already
-		const allChannels = await interaction.member.guild.channels.fetch()
+	
+		const allChannels = await guild.channels.fetch();
 		const numActivePuzzles = allChannels.filter(ch => {
-			return ch.parentId === ACTIVE_PUZZLES_CHANNEL_CATEGORY_ID
+			return ch?.parentId === ACTIVE_PUZZLES_CHANNEL_CATEGORY_ID
 		}).size
 		
 		if (numActivePuzzles >= 5) {
-			await interaction.editReply({content: `⚠️ There are too many puzzles currently active! Try completing some of the existing ones first! 🧠`, ephemeral: true});
+			await interaction.editReply(`⚠️ There are too many puzzles currently active! Try completing some of the existing ones first! 🧠`);
 			return;
 		}
 
@@ -77,6 +83,11 @@ async function newpuzzle(interaction) {
 		const puzzleTitle = await page.locator('.chat--header--title').textContent();
 		await browser.close();
 
+		if (!puzzleTitle) {
+			await interaction.editReply("⚠️ I couldn't find the puzzle title on that page!")
+			return
+		}
+
 		// slugify the title
 		const channelTitle = slugify(puzzleTitle, {
 			remove: /[,._\-'"]/gm,
@@ -86,17 +97,17 @@ async function newpuzzle(interaction) {
 
 		// Check to see if there's another channel already with this title
 		const existingChannel = allChannels.find(ch => {
-			return ch.name === channelTitle;
+			return ch?.name === channelTitle;
 		})
 
 		if (existingChannel) {
-			await interaction.editReply({content: `⚠️ There's already a channel for this puzzle! <#${existingChannel.id}>`, ephemeral: true});
+			await interaction.editReply(`⚠️ There's already a channel for this puzzle! <#${existingChannel.id}>`);
 			return;
 		}
 
 		// make the channel
-		const categoryChannel = await interaction.member.guild.channels.fetch(ACTIVE_PUZZLES_CHANNEL_CATEGORY_ID)
-		const createdChannel = await interaction.member.guild.channels.create({
+		const categoryChannel = await guild.channels.fetch(ACTIVE_PUZZLES_CHANNEL_CATEGORY_ID) as CategoryChannel;
+		const createdChannel = await guild.channels.create({
 			name: channelTitle,
 			type: ChannelType.GuildText,
 			parent: categoryChannel,
@@ -104,7 +115,7 @@ async function newpuzzle(interaction) {
 		});
 
 		// build and send an announcement message
-		const row = new ActionRowBuilder()
+		const row = new ActionRowBuilder<ButtonBuilder>()
 		.addComponents(
 			new ButtonBuilder()
 				.setLabel('Play this crossword!')
@@ -113,41 +124,46 @@ async function newpuzzle(interaction) {
 		)
 
 		const msg = await createdChannel.send({
-			content: `🧩🚨 <@&${PUZZ_WATCHERS_ROLE_ID}> New Puzzle added!\nThanks to <@${interaction.member.id}> for submitting this one! 🤜 🤛`,
+			content: `🧩🚨 <@&${PUZZ_WATCHERS_ROLE_ID}> New Puzzle added!\nThanks to <@${interaction.member?.user.id}> for submitting this one! 🤜 🤛`,
 			components: [row]
 		});
 		await msg.pin();
 
 		// Reply to the user to point them to the new channel
-		await interaction.editReply({content: `<#${createdChannel.id}> has been created! 🎉`, ephemeral: true});
+		await interaction.editReply(`<#${createdChannel.id}> has been created! 🎉`);
 	} catch (e) {
 		console.error("Error in command", e)
-		await interaction.editReply({content: `⚠️ Error running command. Please reach out to my developer`, ephemeral: true});
+		await interaction.editReply(`⚠️ Error running command. Please reach out to my developer`);
 	}
 }
 
-async function closepuzzle(interaction) {
+async function closepuzzle(interaction: ChatInputCommandInteraction) {
 	try {
 		await interaction.deferReply({ephemeral: true});
 
+		// Ensure the `guild` property is accessible
+		if (!interaction.guild) {
+			await interaction.editReply("⚠️ I couldn't find the guild for this interaction!")
+			return
+		}
+		
 		// Check to see if this was invoked in an active puzzle channel
-		const channelSentIn = await interaction.member.guild.channels.fetch(interaction.channelId)
+		const channelSentIn = await interaction.guild.channels.fetch(interaction.channelId) as TextChannel;
 		
 		if (channelSentIn?.parentId !== ACTIVE_PUZZLES_CHANNEL_CATEGORY_ID) {
-			await interaction.editReply({
-				content: "☝️ Tsk tsk! You can't use this command to archive anything but puzzles in the \"Active Puzzles\" category!",
-				ephemeral: true
-			})
+			await interaction.editReply("☝️ Tsk tsk! You can't use this command to archive anything but puzzles in the \"Active Puzzles\" category!")
 			return
 		}
 
 		// Fetch all puzzles in the old puzzles channel
-		const allChannels = await interaction.member.guild.channels.fetch()
+		const allChannels = await interaction.guild.channels.fetch();
 		const oldPuzzles = allChannels.filter(ch => {
-			return ch.parentId === OLD_PUZZLES_CHANNEL_CATEGORY_ID
-		})
+			const isTextBased = ch?.isTextBased();
+			const isPartOfOldChannelCategory = ch && ch.parentId === OLD_PUZZLES_CHANNEL_CATEGORY_ID
+			return isTextBased && isPartOfOldChannelCategory
+		}) as Collection<string, TextChannel>; // This TextChannel cast is safe because we're filtering for text channels with .isTextBased()
 
-		const MAX_OLD_PUZZLES = 45; // Discord has a max of 50 channels, so stay just a few short of that
+		const MAX_OLD_PUZZLES = 2; // Discord has a max of 50 channels, so stay just a few short of that
 		// If we have too many old puzzles and need to start deleting them
 		if (oldPuzzles.size > MAX_OLD_PUZZLES) {
 			// Fetch message counts per channel to identify which to delete
@@ -171,13 +187,19 @@ async function closepuzzle(interaction) {
 			// Delete the puzzles with the smallest number of messages until we get to MAX_OLD_PUZZLES
 			for (let i = 0; i < numPuzzlesToDelete; i++) {
 				const channelIDToDelete = channelsWithIDsAndMessageCounts[i].id;
-				const channelToDelete = await interaction.member.guild.channels.fetch(channelIDToDelete)
-				await channelToDelete.delete();
+				const channelToDelete = await interaction.guild.channels.fetch(channelIDToDelete)
+				if (channelToDelete) {
+					await channelToDelete.delete();
+				}
 			}
 		}
 
 		// Get a screenshot of the puzzle
 		const urlToScreenshot = channelSentIn.topic;
+		if (!urlToScreenshot) {
+			await interaction.editReply("⚠️ I couldn't find the URL for this puzzle!")
+			return
+		}
 		const browser = await chromium.launch();
 		const page = await browser.newPage();
 		await page.setViewportSize({ width: 1280, height: 1080 });
@@ -185,7 +207,8 @@ async function closepuzzle(interaction) {
 		const screenshot = await page.locator('.player--main--left--grid').screenshot();
 		const numPlayers = await page.locator('.dot').count() - 2; // minus two because of the bot being a player when visiting this page (once on create, once on close)
 		await page.waitForTimeout(1000) // wait for the clock to update
-		const solvingTime = (await page.locator('.clock').textContent()).replaceAll("(",'').replaceAll(")","")
+		const solvingTimeRaw = await page.locator('.clock').textContent()
+		const solvingTime = solvingTimeRaw?.replaceAll("(",'').replaceAll(")","") || 'Unknown';
 		await browser.close();
 
 		const attachment = new AttachmentBuilder(screenshot, {name: `${channelSentIn.name}.png`});
@@ -197,7 +220,7 @@ async function closepuzzle(interaction) {
 				{ name: '🕔 Time to Complete', value: `${timeDiff}` },
 				{ name: '🤔 Time Spent Solving', value: `${solvingTime}` },
 				{ name: '👥 Number of Players', value: `${numPlayers}` },
-				{ name: '👨‍💻 Finished by', value: `<@${interaction.member.id}>` },
+				{ name: '👨‍💻 Finished by', value: `<@${interaction.member?.user.id}>` },
 			)
 
 		await channelSentIn.send({
@@ -209,9 +232,9 @@ async function closepuzzle(interaction) {
 		await channelSentIn.setParent(OLD_PUZZLES_CHANNEL_CATEGORY_ID)
 
 		// Reply with a success message
-		await interaction.editReply({content: `You closed <#${channelSentIn.id}>`, ephemeral: true});
+		await interaction.editReply(`You closed <#${channelSentIn.id}>`);
 	} catch (e) {
 		console.error("Error in command", e)
-		await interaction.editReply({content: `⚠️ Error running command. Please reach out to my developer`, ephemeral: true});
+		await interaction.editReply(`⚠️ Error running command. Please reach out to my developer`);
 	}
 }
